fix dhcp build on OSX
